requirements.txt
python-telegram-bot>=20.8
google-generativeai>=0.3.2